stock-agent-v2/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
akshare>=1.14.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
tenacity>=8.2.0
